+++ R/installConda.R
@@ -17,55 +17,23 @@ NULL
 #' 
 #' @examples
 #' \dontrun{
-#' # Create a new SyncroSim Session
-#' mySession <- session()
-#' 
-#' # Install Conda for the given SyncroSim session
-#' installConda(mySession)
+#' # Install Conda for the default SyncroSim session
+#' installConda()
 #' }
 #' 
 #' @export
-setGeneric("installConda", function(session = NULL) standardGeneric("installConda"))
-
-#' @rdname installConda
-setMethod("installConda", signature(session = "character"), function(session) {
-  return(SyncroSimNotFound(session))
-})
-
-#' @rdname installConda
-setMethod("installConda", signature(session = "missingOrNULL"), function(session) {
-  session <- .session()
-  return(installConda(session))
-})
-
-#' @rdname installConda
-setMethod("installConda", signature(session = "Session"), function(session) {
+installConda <- function() {
+  
   success <- FALSE
-
+  message("Setting Conda filepath to the default installation.")
   args <- list(conda = NULL, install = NULL)
+
+  message("Running Conda Installer.  Please wait...")
+  session <- .session()
+  tt <- command(args, session)
+
   
-  installRunning <- TRUE
-  time <- 0
-  
-  while (installRunning & time < 600) {
-    tt <- command(args, session)
-    if (endsWith(tt[2], "Running Conda Installer.  Please wait...")){
-      message("Running Conda Installer.  Please wait...")
-    } else {
-      installRunning <- FALSE
-    }
-    
-    if (length(tt) > 2) {
-      installRunning <- FALSE
-      message(tt[3])
-    }
-    
-    addTime <- 5
-    Sys.sleep(addTime)
-    time <- time + addTime
-  }
-  
-  if (tt[3] == "saved") {
+  if (tt[3] == "Saved") {
     success <- TRUE
     tt <- paste0("Miniconda successfully installed")
   } else if (tt[1] == "Conda already installed at that location"){
@@ -74,5 +42,6 @@ setMethod("installConda", signature(session = "Session"), function(session) {
   }
   
   message(tt)
+  
   return(invisible(success))
-})
+}